% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_constants.R
\docType{data}
\name{ANALYTE_PRIORS}
\alias{ANALYTE_PRIORS}
\title{Predefined hyperparameter priors for specific analytes}
\format{
An object of class \code{list} of length 5.
}
\usage{
ANALYTE_PRIORS
}
\description{
Predefined hyperparameter priors for specific analytes
}
\keyword{datasets}
