% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_constants.R
\docType{data}
\name{ANALYTE_SYNONYMS}
\alias{ANALYTE_SYNONYMS}
\title{Analyte name synonym mapping}
\format{
An object of class \code{list} of length 4.
}
\usage{
ANALYTE_SYNONYMS
}
\description{
Analyte name synonym mapping
}
\keyword{datasets}
