% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biovar_ported.R
\name{calculate_cv}
\alias{calculate_cv}
\title{Calculate a coefficient of variation (%) from SD and mean.}
\usage{
calculate_cv(sd, mean)
}
\arguments{
\item{sd}{Numeric vector of standard deviations.}

\item{mean}{Numeric vector of means.}
}
\value{
Numeric vector of CVs in percent.
}
\description{
Calculate a coefficient of variation (%) from SD and mean.
}
\keyword{internal}
