% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biovar_ported.R
\name{calculate_trace_and_posterior_plotting_data}
\alias{calculate_trace_and_posterior_plotting_data}
\title{Calculate Trace Plot and Posterior Distribution Plotting data}
\usage{
calculate_trace_and_posterior_plotting_data(trace_fit, log_transformed)
}
\arguments{
\item{trace_fit}{A \code{data.frame} generated from
\code{rstan::traceplot(...)$data}.}

\item{log_transformed}{A \code{logical} value. Set to
\code{TRUE} if the multiplicative model is utilized.}
}
\value{
A \code{list} object containing three \code{data.table} objects:
\itemize{
   \item \code{univariate_parameters}: Draws for single
     parameters (beta, dfs, CVs).
   \item \code{homeostatic_set_points}: Draws for
     subject-specific means (mu_p(i)).
   \item \code{subject_specific_cvs}: Draws for
     subject-specific CVs (CV_p(i)).
}
}
\description{
Calculate Trace Plot and Posterior Distribution Plotting data
}
