% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_d3_plots.R
\name{clean_finite_values}
\alias{clean_finite_values}
\title{Remove Non-Finite Values and Optionally Thin}
\usage{
clean_finite_values(x, max_length = 5000L)
}
\arguments{
\item{x}{Numeric vector.}

\item{max_length}{Maximum length of the returned vector.}
}
\value{
A numeric vector with only finite values, thinned
  if necessary.
}
\description{
Remove Non-Finite Values and Optionally Thin
}
\keyword{internal}
