% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glassAnalysisProgress.R
\name{closeGlassProgress}
\alias{closeGlassProgress}
\title{Close the Glass Analysis Progress overlay}
\usage{
closeGlassProgress(delay = 800, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{delay}{Milliseconds to wait before fade-out (lets user see 100%)}

\item{session}{Shiny session}
}
\description{
Close the Glass Analysis Progress overlay
}
