% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_constants.R
\docType{data}
\name{COLUMN_SELECTORS}
\alias{COLUMN_SELECTORS}
\title{Default names for expected columns to exist in uploaded data}
\format{
An object of class \code{character} of length 10.
}
\usage{
COLUMN_SELECTORS
}
\description{
Default names for expected columns to exist in uploaded data
}
\keyword{datasets}
