% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_plotting_functions.R
\name{compute_iteration_breaks}
\alias{compute_iteration_breaks}
\title{Compute iteration-axis breaks based on the total
number of post-warmup iterations.}
\usage{
compute_iteration_breaks(iterations)
}
\arguments{
\item{iterations}{Numeric vector of iteration IDs.}
}
\value{
A \code{list} with elements \code{breaks}
  and \code{shift} (the minimum iteration ID).
}
\description{
Compute iteration-axis breaks based on the total
number of post-warmup iterations.
}
\keyword{internal}
