% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_d3_plots.R
\name{compute_population_standard_rcv}
\alias{compute_population_standard_rcv}
\title{Compute Population-Level Standard RCV}
\usage{
compute_population_standard_rcv(cv_analytical, cv_individual)
}
\arguments{
\item{cv_analytical}{Numeric. Analytical CV (%).}

\item{cv_individual}{Numeric. Individual CV (%).}
}
\value{
A list with \code{upper} and \code{lower} RCV.
}
\description{
Compute Population-Level Standard RCV
}
\keyword{internal}
