% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_documentation.R
\name{.doc_info_box}
\alias{.doc_info_box}
\title{Build an info box with a coloured left border}
\usage{
.doc_info_box(..., colour = "#605CA8")
}
\description{
Build an info box with a coloured left border
}
\keyword{internal}
