% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_documentation.R
\name{.doc_section}
\alias{.doc_section}
\title{Build a documentation section with header and body}
\usage{
.doc_section(..., header = NULL)
}
\description{
Build a documentation section with header and body
}
\keyword{internal}
