% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_analysis.R
\name{format_bv_value}
\alias{format_bv_value}
\title{Format a numeric value for display in BV tables}
\usage{
format_bv_value(x)
}
\arguments{
\item{x}{Numeric vector of values to format.}
}
\value{
Character vector of formatted values.
}
\description{
Rounds to 2 decimal places. If the value is positive but less than 0.005
(i.e., would round to 0.00), it is displayed as "< 0.01" instead.
}
\keyword{internal}
