% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biovar_ported.R
\name{get_percent_suffix}
\alias{get_percent_suffix}
\title{Return \code{" \%"} for CV parameters, \code{""} for
others.}
\usage{
get_percent_suffix(parameter_name)
}
\arguments{
\item{parameter_name}{A character string (plotmath
label).}
}
\value{
A single-element character string.
}
\description{
Return \code{" \%"} for CV parameters, \code{""} for
others.
}
\keyword{internal}
