% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biovar_ported.R
\name{get_univariate_parameter_order}
\alias{get_univariate_parameter_order}
\title{Canonical order for univariate parameter labels.}
\usage{
get_univariate_parameter_order()
}
\value{
A character vector of parameter label strings.
}
\description{
Canonical order for univariate parameter labels.
}
\keyword{internal}
