% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glassLayout.R
\name{glassCol}
\alias{glassCol}
\title{Glass Grid Column}
\usage{
glassCol(
  width,
  ...,
  label = NULL,
  label_icon = NULL,
  label_position = "left",
  label_underline = FALSE
)
}
\arguments{
\item{width}{Integer between 1 and 12. 12 is full width, 6 is half width.}

\item{...}{Content.}

\item{label}{Optional text label for the column header. Default NULL.}

\item{label_icon}{Optional icon for the column header. Default NULL.}

\item{label_position}{Position of the label: "left", "center", or "right".
Default "left".}

\item{label_underline}{Logical. If TRUE, adds an underline to the label.
Default FALSE.}
}
\description{
A column within a `glassRow`.
}
