% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glassD3Plot.R
\name{glassD3PlotOutput}
\alias{glassD3PlotOutput}
\title{Glass D3 Plot Container (UI)}
\usage{
glassD3PlotOutput(outputId, width = "100\%", height = "600px")
}
\arguments{
\item{outputId}{A unique ID for this chart container.}

\item{width}{CSS width. Defaults to \code{"100\%"}.}

\item{height}{CSS height. Defaults to \code{"600px"}.}
}
\value{
An \code{htmltools::tags$div}.
}
\description{
Creates a \code{<div>} container in which the D3 plot will be rendered.
}
