% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glassDropdown.R
\name{glassDropdown}
\alias{glassDropdown}
\title{Glass Dropdown Input}
\usage{
glassDropdown(
  inputId,
  label = NULL,
  label_icon = NULL,
  help_text = NULL,
  choices,
  selected = NULL,
  disabled = FALSE,
  color = "purple",
  width = "100\%"
)
}
\arguments{
\item{inputId}{The input slot that will be used to access the value.}

\item{label}{Optional display label text.}

\item{label_icon}{Optional \code{icon()} for the label.}

\item{help_text}{Optional tooltip text to display on hover.}

\item{choices}{List of values to select from.
Can be a named vector c("Label" = "val").}

\item{selected}{The initially selected value
(must match the value, not the label).}

\item{disabled}{Logical. If \code{TRUE}, the dropdown is disabled on load.}

\item{color}{Color variant: "purple" (default) or "green".}

\item{width}{The width of the input.}
}
\description{
Glass Dropdown Input
}
