% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glassFileInput.R
\name{glassFileInput}
\alias{glassFileInput}
\title{Glass File Input}
\usage{
glassFileInput(
  inputId,
  label,
  label_icon = NULL,
  accept = NULL,
  button_label = "Browse...",
  button_label_icon = NULL,
  placeholder = "No file selected",
  tooltip_empty = "Please select a file",
  tooltip_filled = "File selected",
  help_text = NULL,
  width = "100\%"
)
}
\arguments{
\item{inputId}{The input slot that will be used to access the value.}

\item{label}{Display label for the control, or NULL for no label.}

\item{label_icon}{Optional icon() to place next to the label.}

\item{accept}{A character vector of MIME types or file extensions
(e.g. c(".csv", ".xlsx")).}

\item{button_label}{The text on the browse button.}

\item{button_label_icon}{An shiny::icon() for the browse button.}

\item{placeholder}{The text to display before a file is selected.}

\item{tooltip_empty}{Text to show when hovering
the text area before upload.}

\item{tooltip_filled}{Text to show when hovering
the text area after file selection.}

\item{help_text}{Text to show when hovering the help-icon.}

\item{width}{The width of the input.}
}
\description{
Glass File Input
}
