% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glassLayout.R
\name{glassFlyoutItem}
\alias{glassFlyoutItem}
\title{Glass Flyout Menu Item}
\usage{
glassFlyoutItem(tabName, icon, label, tabset_id = NULL, tab_value = NULL)
}
\arguments{
\item{tabName}{The ID of the page/view to navigate to.}

\item{icon}{The icon to display.}

\item{label}{The text label.}
}
\description{
Acts in the same way as glassTabsetPanel's tabPanel.
}
