% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glassHeaderIndicator.R
\name{glassHeaderIndicator}
\alias{glassHeaderIndicator}
\title{Glass Header Indicator}
\usage{
glassHeaderIndicator(
  inputId,
  type = c("filter", "reference", "diagnostics"),
  icon_name = NULL,
  tooltip_text = "",
  visible = FALSE
)
}
\arguments{
\item{inputId}{Unique ID.}

\item{type}{Type of indicator: "filter" (default),
"reference", or "diagnostics".}

\item{tooltip_text}{Initial text to display on hover.}

\item{visible}{Logical. Should it be visible initially?
Defaults to FALSE.}
}
\description{
A versatile indicator for the dashboard header. Give interactive feedback
to the user about what is going on in the app, so they feel informed and in control.
}
