% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glassInput.R
\name{glassNumericInput}
\alias{glassNumericInput}
\title{Glass Numeric Input}
\usage{
glassNumericInput(
  inputId,
  label,
  value,
  min = NA,
  max = NA,
  step = NA,
  label_icon = NULL,
  help_text = NULL,
  placeholder = "",
  tooltip_empty = "Enter number",
  tooltip_filled = "Number entered",
  width = "100\%",
  unit = NULL,
  accept = c(-Inf, Inf),
  warning_unacceptable = "Value is outside acceptable range",
  color = "green",
  disabled = FALSE
)
}
\arguments{
\item{inputId}{The input slot.}

\item{label}{Display label.}

\item{value}{Initial value.}

\item{min}{Minimum allowed value (for HTML attribute).}

\item{max}{Maximum allowed value (for HTML attribute).}

\item{step}{Step size.}

\item{label_icon}{Optional \code{icon()}.}

\item{help_text}{Text to show on help icon hover.}

\item{placeholder}{Placeholder text.}

\item{tooltip_empty}{Text on hover when empty.}

\item{tooltip_filled}{Text on hover when filled.}

\item{width}{Width of input.}

\item{unit}{Optional unit string (e.g. "mg/L").}

\item{accept}{A numeric vector of length 2 \code{c(min, max)}
defining the acceptable range.}

\item{warning_unacceptable}{Text to display if value is outside
\code{accept} range.}

\item{color}{Color variant: "green" (default) or "purple".}

\item{disabled}{Logical.}
}
\description{
Glass Numeric Input
}
