% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glassLayout.R
\name{glassPage}
\alias{glassPage}
\title{Glass Layout Shell}
\usage{
glassPage(title, branding, sidebar, header_items = NULL, ...)
}
\arguments{
\item{title}{App Title. Displayed in header.}

\item{branding}{UI element for top-left. Typically a logo image.
Can also include text, but that is less common.}

\item{sidebar}{UI element from `glassSidebar`. The navigation sidebar.
Vertical icon-based navigation between `glassRoute` pages.}

\item{header_items}{UI elements for header. Typically status indicators,
user profile menu, etc.}

\item{...}{`glassRouter` content. The actual pages of the app.
Corresponds to `tabName` in `glassNavItem`.}
}
\description{
Glass Layout Shell
}
