% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glassResultCard.R
\name{glassResultCard}
\alias{glassResultCard}
\title{Glass Result Card}
\usage{
glassResultCard(
  inputId,
  title,
  ...,
  toolbar = NULL,
  icon = NULL,
  footer = NULL,
  attached = FALSE,
  width = "100\%"
)
}
\arguments{
\item{inputId}{The id of the card.}

\item{title}{The title text.}

\item{...}{Content (The Table or Plot).}

\item{toolbar}{UI elements to place in the header (e.g. buttons).}

\item{icon}{An optional icon().}

\item{footer}{Optional footer text or UI.}

\item{attached}{Logical. If TRUE, removes top border-radius to blend with a boxed tabset above.}

\item{width}{Width of the card.}
}
\description{
A container optimized for displaying tables and plots.
Includes a header toolbar for action buttons (Calculate, Download).
}
