% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glassLayout.R
\name{glassRoute}
\alias{glassRoute}
\title{Glass Page (Route)}
\usage{
glassRoute(title, ...)
}
\arguments{
\item{title}{The ID used by navigation (must match tabName in navItem).}

\item{...}{Content.}
}
\description{
A container for a single "tab" of content. Replaces tabPanel.
}
