% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glassLayout.R
\name{glassSidebar}
\alias{glassSidebar}
\title{Glass Sidebar Container}
\usage{
glassSidebar(inputId, ...)
}
\arguments{
\item{inputId}{Shiny input ID to track navigation state.}

\item{...}{`glassNavItem` elements.}
}
\description{
Glass Sidebar Container
}
