% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glassSlider.R
\name{glassSlider}
\alias{glassSlider}
\title{Glass Slider Input}
\usage{
glassSlider(
  inputId,
  label = NULL,
  label_icon = NULL,
  help_text = NULL,
  choices,
  selected = NULL,
  unit = "",
  color = "purple",
  width = "100\%"
)
}
\arguments{
\item{inputId}{The input slot that will be used to access the value.}

\item{label}{Optional label for the slider input.}

\item{label_icon}{Optional \code{icon()} to display next to the label}

\item{help_text}{Optional tooltip text to display on hover.}

\item{choices}{Vector of values (numbers or strings).}

\item{selected}{The initially selected value.}

\item{unit}{Optional string to append to values (e.g. " dpi").}

\item{color}{Color variant: "purple" (default) or "green".}

\item{width}{The width of the input, e.g., '100%'.}
}
\description{
Glass Slider Input
}
