% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glassLayout.R
\name{glassSpacerV}
\alias{glassSpacerV}
\title{Glass Vertical Spacer}
\usage{
glassSpacerV()
}
\value{
An HTML div element with the vertical spacer styling.
}
\description{
A subtle gradient bar to visually connect side-by-side cards in a row.
Place between glassCol elements within a glassRow.
}
