% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glassTabsetPanel.R
\name{glassTabPanel}
\alias{glassTabPanel}
\title{Glass Tab Panel}
\usage{
glassTabPanel(title, ..., value = title, icon = NULL)
}
\arguments{
\item{title}{A \code{character} string. The title of the glassTabPanel.}

\item{...}{HTML input.}

\item{value}{A \code{character} string. The id of the glassTabPanel.}

\item{icon}{A \code{icon(...)} appearing just before \code{title}.}
}
\description{
Glass Tab Panel
}
