% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glassInput.R
\name{glassTextInput}
\alias{glassTextInput}
\title{Glass Text Input}
\usage{
glassTextInput(
  inputId,
  label,
  value = "",
  label_icon = NULL,
  help_text = NULL,
  placeholder = "",
  tooltip_empty = "Enter text",
  tooltip_filled = "Text entered",
  width = "100\%",
  unit = NULL,
  color = "green",
  disabled = FALSE
)
}
\arguments{
\item{inputId}{The input slot that will be used to access the value.}

\item{label}{Display label for the control, or NULL for no label.}

\item{value}{Initial value.}

\item{label_icon}{Optional \code{icon()} to place next to the label.}

\item{help_text}{Text to show when hovering the help-icon (top right).}

\item{placeholder}{The placeholder text inside the input.}

\item{tooltip_empty}{Text to show on hover when input is empty.}

\item{tooltip_filled}{Text to show on hover when input has content.}

\item{width}{The width of the input (e.g., "auto").}

\item{unit}{Optional text to display on the right side
(e.g. "kg", "mmol/mol").}

\item{color}{Color variant: "green" (default) or "purple".}

\item{disabled}{Logical. Should it be disabled on load?}
}
\description{
Glass Text Input
}
