% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{guess_columns}
\alias{guess_columns}
\title{Guess column mappings from data column names}
\usage{
guess_columns(col_names)
}
\arguments{
\item{col_names}{Character vector of column names}
}
\value{
Named list of guessed column mappings
}
\description{
Guess column mappings from data column names
}
