% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_stan_output.R
\name{lognormal_sd_to_cv}
\alias{lognormal_sd_to_cv}
\title{Convert a log-normal standard deviation to a CV (%).}
\usage{
lognormal_sd_to_cv(log_sd)
}
\arguments{
\item{log_sd}{Numeric vector of log-scale standard deviations.}
}
\value{
Numeric vector of CVs in percent.
}
\description{
Convert a log-normal standard deviation to a CV (%).
}
\keyword{internal}
