% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing_stan_output.R
\name{logt_scale_to_cv}
\alias{logt_scale_to_cv}
\title{Convert a log-Student-t scale parameter to a CV (%).}
\usage{
logt_scale_to_cv(scale, df)
}
\arguments{
\item{scale}{Numeric vector of scale parameters.}

\item{df}{Numeric vector of degrees-of-freedom parameters.}
}
\value{
Numeric vector of CVs in percent.
}
\description{
For a log-t distributed variable, the variance is equivalent to a
log-normal with \code{sd = sqrt(scale^2 * df / (df - 2))}.
}
\keyword{internal}
