% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_column_map.R
\name{mod_column_map_server}
\alias{mod_column_map_server}
\title{Column Mapping Module Server}
\usage{
mod_column_map_server(id, uploaded_data)
}
\arguments{
\item{id}{Namespace ID for the module}

\item{uploaded_data}{Reactive containing uploaded data}
}
\value{
List of reactive values for column selections
}
\description{
Column Mapping Module Server
}
