% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_column_map.R
\name{mod_column_map_ui}
\alias{mod_column_map_ui}
\title{Column Mapping Module UI}
\usage{
mod_column_map_ui(id)
}
\arguments{
\item{id}{Namespace ID for the module}
}
\value{
Shiny UI elements
}
\description{
Column Mapping Module UI
}
