% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_data_exploration.R
\name{mod_data_exploration_server}
\alias{mod_data_exploration_server}
\title{Data Exploration Module Server}
\usage{
mod_data_exploration_server(id, app_state)
}
\arguments{
\item{id}{Namespace ID for the module}

\item{app_state}{reactiveValues object containing shared application state}
}
\description{
Data Exploration Module Server
}
