% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_data_exploration.R
\name{mod_data_exploration_ui}
\alias{mod_data_exploration_ui}
\title{Data Exploration Module UI}
\usage{
mod_data_exploration_ui(id)
}
\arguments{
\item{id}{Namespace ID for the module}
}
\value{
Shiny UI elements for the data exploration page
}
\description{
Data Exploration Module UI
}
