% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_documentation.R
\name{mod_documentation_server}
\alias{mod_documentation_server}
\title{Documentation Module Server}
\usage{
mod_documentation_server(id)
}
\arguments{
\item{id}{Namespace ID for the module}
}
\description{
Documentation Module Server
}
