% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_documentation.R
\name{mod_documentation_ui}
\alias{mod_documentation_ui}
\title{Documentation Module UI}
\usage{
mod_documentation_ui(id)
}
\arguments{
\item{id}{Namespace ID for the module}
}
\value{
Shiny UI elements for the documentation page
}
\description{
Documentation Module UI
}
