% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_filter.R
\name{mod_filter_server}
\alias{mod_filter_server}
\title{Filter & Name Module Server}
\usage{
mod_filter_server(id, uploaded_data, col_selections)
}
\arguments{
\item{id}{Namespace ID for the module}

\item{uploaded_data}{Reactive containing uploaded data frame}

\item{col_selections}{List of reactive column selections from
column mapping module}
}
\value{
List of reactive values for filters and names
}
\description{
Server logic for the filter and name module. Creates dynamic filter inputs
based on uploaded data and column selections, and manages user-defined names.
}
