% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_filter.R
\name{mod_filter_ui}
\alias{mod_filter_ui}
\title{Filter & Name Module UI}
\usage{
mod_filter_ui(id)
}
\arguments{
\item{id}{Namespace ID for the module}
}
\value{
Shiny UI elements
}
\description{
Creates a card interface for filtering data by various categories and
assigning custom names to groups.
}
