% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_hyperparameters.R
\name{mod_hyperparameters_server}
\alias{mod_hyperparameters_server}
\title{Hyperparameters Module Server}
\usage{
mod_hyperparameters_server(id, filter_values = NULL)
}
\arguments{
\item{id}{Namespace ID for the module}

\item{filter_values}{Reactive list from filter module
(contains selected analyte)}
}
\value{
List of reactive values for hyperparameters
}
\description{
Hyperparameters Module Server
}
