% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_hyperparameters.R
\name{mod_hyperparameters_ui}
\alias{mod_hyperparameters_ui}
\title{Hyperparameters Module UI}
\usage{
mod_hyperparameters_ui(id)
}
\arguments{
\item{id}{Namespace ID for the module}
}
\value{
Shiny UI elements
}
\description{
Hyperparameters Module UI
}
