% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_model_ntt.R
\name{mod_model_ntt_ui}
\alias{mod_model_ntt_ui}
\title{Model 1 (NTT) Module UI}
\usage{
mod_model_ntt_ui(id)
}
\arguments{
\item{id}{Namespace ID for the module}
}
\value{
Shiny UI elements for the NTT model results page
}
\description{
Model 1 (NTT) Module UI
}
