% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_model_nttdfgam.R
\name{mod_model_nttdfgam_server}
\alias{mod_model_nttdfgam_server}
\title{Model 2 (NTTDFGAM) Module Server}
\usage{
mod_model_nttdfgam_server(id, app_state)
}
\arguments{
\item{id}{Namespace ID for the module}

\item{app_state}{reactiveValues object containing shared application state}
}
\description{
Model 2 (NTTDFGAM) Module Server
}
