% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_model_nttdfgam.R
\name{mod_model_nttdfgam_ui}
\alias{mod_model_nttdfgam_ui}
\title{Model 2 (NTTDFGAM) Module UI}
\usage{
mod_model_nttdfgam_ui(id)
}
\arguments{
\item{id}{Namespace ID for the module}
}
\value{
Shiny UI elements for the NTTDFGAM model results page
}
\description{
Model 2 (NTTDFGAM) Module UI
}
