% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_setup.R
\name{mod_setup_server}
\alias{mod_setup_server}
\title{Setup Module Server}
\usage{
mod_setup_server(id, app_state)
}
\arguments{
\item{id}{Namespace ID for the module}

\item{app_state}{reactiveValues object containing shared application state}
}
\description{
Setup Module Server
}
