% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_setup.R
\name{mod_setup_ui}
\alias{mod_setup_ui}
\title{Setup Module UI}
\usage{
mod_setup_ui(id)
}
\arguments{
\item{id}{Namespace ID for the module}
}
\value{
Shiny UI elements for the setup wizard
}
\description{
Setup Module UI
}
