% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_upload.R
\name{mod_upload_server}
\alias{mod_upload_server}
\title{Data Upload Module Server}
\usage{
mod_upload_server(id)
}
\arguments{
\item{id}{Namespace ID for the module}
}
\value{
Reactive value containing uploaded data
}
\description{
Data Upload Module Server
}
