% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_upload.R
\name{mod_upload_ui}
\alias{mod_upload_ui}
\title{Data Upload Module UI}
\usage{
mod_upload_ui(id)
}
\arguments{
\item{id}{Namespace ID for the module}
}
\value{
Shiny UI elements
}
\description{
Data Upload Module UI
}
