% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{parse_stan_warnings}
\alias{parse_stan_warnings}
\title{Parse Stan warnings and provide user-friendly advice}
\usage{
parse_stan_warnings(
  warnings,
  current_iter,
  current_adapt_delta,
  current_max_treedepth
)
}
\arguments{
\item{warnings}{Character vector of warning messages from Stan}

\item{current_iter}{Current number of iterations}

\item{current_adapt_delta}{Current adapt_delta value}

\item{current_max_treedepth}{Current max_treedepth value}
}
\value{
List of HTML tags with advice for each warning type
}
\description{
Parse Stan warnings and provide user-friendly advice
}
