% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_for_stan_modeling.R
\name{pernille_prepares_for_stan_modeling}
\alias{pernille_prepares_for_stan_modeling}
\title{Prepare data for Stan modeling}
\usage{
pernille_prepares_for_stan_modeling(
  data,
  hypers,
  hyper_strength,
  log_transformed = FALSE
)
}
\arguments{
\item{data}{Data table with biological variation data}

\item{hypers}{List of hyperparameters}

\item{hyper_strength}{Numeric strength of hyperparameters}

\item{log_transformed}{Logical indicating if log transformation
should be applied}
}
\value{
List of prepared data for Stan
}
\description{
Prepare data for Stan modeling
}
