% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_d3_plots.R
\name{prepare_cvi_vs_concentration_d3}
\alias{prepare_cvi_vs_concentration_d3}
\title{Prepare Data for CVI vs. Concentration D3 Chart}
\usage{
prepare_cvi_vs_concentration_d3(
  processed_output,
  title = paste("Subject-Specific CVs", "vs. Concentration"),
  subtitle = NULL
)
}
\arguments{
\item{processed_output}{A list returned by
\code{process_stan_output()}.}

\item{title}{Chart title.}

\item{subtitle}{Chart subtitle.}
}
\value{
A list with \code{plot_type}, \code{data},
  \code{params}.
}
\description{
Prepare Data for CVI vs. Concentration D3 Chart
}
