% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_d3_plots.R
\name{prepare_cvi_vs_rcv_d3}
\alias{prepare_cvi_vs_rcv_d3}
\title{Prepare Data for CVI vs. RCV D3 Chart}
\usage{
prepare_cvi_vs_rcv_d3(
  processed_output,
  use_log_rcv = FALSE,
  title = "Subject-Specific RCVs with Uncertainty",
  subtitle = paste("Incorporated from", "CV Credible Intervals")
)
}
\arguments{
\item{processed_output}{A list returned by
\code{process_stan_output()}.}

\item{use_log_rcv}{Logical; use log-based RCV formula?}

\item{title}{Chart title.}

\item{subtitle}{Chart subtitle.}
}
\value{
A list with \code{plot_type}, \code{data},
  \code{params}.
}
\description{
Prepare Data for CVI vs. RCV D3 Chart
}
