% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glassFileInput.R
\name{processGlassFile}
\alias{processGlassFile}
\title{Helper to process Glass File in R}
\usage{
processGlassFile(input_data)
}
\arguments{
\item{input_data}{The value from input$yourId}
}
\description{
Helper to process Glass File in R
}
