% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glassToast.R
\name{removeAllGlassToasts}
\alias{removeAllGlassToasts}
\title{Remove All Glass Toast Notifications}
\usage{
removeAllGlassToasts(session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{session}{The Shiny session object.}
}
\description{
Remove all currently visible toast notifications.
}
