% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glassToast.R
\name{removeGlassToast}
\alias{removeGlassToast}
\title{Remove a Glass Toast Notification}
\usage{
removeGlassToast(id, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{id}{The ID of the toast to remove.}

\item{session}{The Shiny session object.}
}
\description{
Remove a specific toast notification by its ID.
}
