% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_d3_plots.R
\name{sample_gamma_ms}
\alias{sample_gamma_ms}
\title{Sample from a Gamma Distribution (Mean / SD Form)}
\usage{
sample_gamma_ms(n, mean, sd)
}
\arguments{
\item{n}{Number of samples.}

\item{mean}{Mean of the gamma distribution.}

\item{sd}{Standard deviation of the gamma distribution.}
}
\value{
Numeric vector of length \code{n}.
}
\description{
Parameterises the gamma distribution by its mean and
standard deviation, then draws samples.
}
\keyword{internal}
