% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_d3_plots.R
\name{transform_prior_to_user_scale}
\alias{transform_prior_to_user_scale}
\title{Transform Prior Samples to User-Friendly CV Scale}
\usage{
transform_prior_to_user_scale(samples, log_transformed, beta)
}
\arguments{
\item{samples}{Named list of prior sample vectors (from
\code{draw_prior_samples()}).}

\item{log_transformed}{Logical.}

\item{beta}{Numeric. Expected beta value.}
}
\value{
A named list of user-scale sample vectors.
}
\description{
Transform Prior Samples to User-Friendly CV Scale
}
\keyword{internal}
