% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glassButton.R
\name{triggerGlassDownload}
\alias{triggerGlassDownload}
\title{Trigger Glass Download}
\usage{
triggerGlassDownload(session, inputId, url, filename = NULL)
}
\arguments{
\item{session}{Shiny session object.}

\item{inputId}{The ID of the glassDownloadButton.}

\item{url}{The URL of the file to download (path to www or temp file).}

\item{filename}{Optional. The name to save the file as.}
}
\description{
Sends a custom message to the glassDownloadButton to initiate a download.
Useful when generating files dynamically in an observeEvent.
}
