% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glassButton.R
\name{updateGlassButton}
\alias{updateGlassButton}
\title{Update Glass Button}
\usage{
updateGlassButton(
  session,
  inputId,
  label = NULL,
  icon = NULL,
  color = NULL,
  disabled = NULL,
  urgent = NULL,
  urgent_text = NULL
)
}
\arguments{
\item{session}{A valid Shiny session object.}

\item{inputId}{The input slot that will be used to access the value.}

\item{label}{The label to display.}

\item{icon}{An optional icon() to appear in the white section.}

\item{color}{Update the color style ("purple", "green", "white").}

\item{disabled}{Logical. Should the button be disabled?}

\item{urgent}{Logical. Should the urgent animation be shown?}

\item{urgent_text}{Character. Update the urgent message text.}
}
\description{
Update Glass Button
}
