% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glassCard.R
\name{updateGlassCard}
\alias{updateGlassCard}
\title{Update Glass Dashboard Card}
\usage{
updateGlassCard(
  session,
  inputId,
  title = NULL,
  collapsed = NULL,
  disabled = NULL
)
}
\arguments{
\item{session}{The Shiny session object.}

\item{inputId}{The id of the card.}

\item{title}{Optional. New title text.}

\item{collapsed}{Optional. TRUE/FALSE.}

\item{disabled}{Optional. TRUE/FALSE.}
}
\description{
Update the title, collapse state, or disabled state of a glassCard.
NEED FIX: Should be able to update content, toolbar, icon, and
collapsible state as well in the future.
}
