% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glassButton.R
\name{updateGlassDownloadButton}
\alias{updateGlassDownloadButton}
\title{Update Glass Download Button}
\usage{
updateGlassDownloadButton(
  session,
  outputId,
  label = NULL,
  icon = NULL,
  color = NULL,
  disabled = NULL
)
}
\arguments{
\item{session}{A valid Shiny session object.}

\item{outputId}{The output slot
(which is also the input ID for the button state).}

\item{label}{The label to display.}

\item{icon}{An optional icon() to appear in the white section.}

\item{color}{Update the color style ("purple", "green", "white").}

\item{disabled}{Logical. Should the button be disabled?}
}
\description{
Alias for updateGlassButton to clarify usage with download buttons.
}
