% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glassDropdown.R
\name{updateGlassDropdown}
\alias{updateGlassDropdown}
\title{Update Glass Dropdown}
\usage{
updateGlassDropdown(
  session,
  inputId,
  choices = NULL,
  selected = NULL,
  disabled = NULL,
  color = NULL
)
}
\arguments{
\item{session}{The Shiny session object.}

\item{inputId}{The id of the input object.}

\item{choices}{Optional. New list of choices (can be a named vector).}

\item{selected}{Optional. The value to select.}

\item{disabled}{Optional. TRUE to disable, FALSE to enable.}

\item{color}{Optional. Color variant: "green" or "purple".}
}
\description{
Updates the choices, selection, or disabled state of a glassDropdown on the client.
}
