% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glassHeaderIndicator.R
\name{updateGlassHeaderIndicator}
\alias{updateGlassHeaderIndicator}
\title{Update Glass Header Indicator}
\usage{
updateGlassHeaderIndicator(
  session,
  inputId,
  visible = NULL,
  tooltip_text = NULL,
  status = NULL,
  icon_name = NULL
)
}
\arguments{
\item{session}{Shiny session.}

\item{inputId}{ID of the indicator.}

\item{visible}{Logical. Show or hide?}

\item{tooltip_text}{String. New text for tooltip.}

\item{status}{String. Optional status class
("success", "warning", "error", "info").
Only applies to 'diagnostics' type to change color.}

\item{icon_name}{String. Optional. Name of a new icon to set
(e.g. "triangle-exclamation").
Mostly for 'diagnostics' type.}
}
\description{
Updates the state, text, and styling of the header indicator.
}
