% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glassInput.R
\name{updateGlassNumericInput}
\alias{updateGlassNumericInput}
\title{Update Glass Numeric Input}
\usage{
updateGlassNumericInput(
  session,
  inputId,
  label = NULL,
  value = NULL,
  accept = NULL,
  disabled = NULL,
  color = NULL
)
}
\arguments{
\item{session}{A valid shiny session object}

\item{inputId}{The local inputId (not the full ID)}

\item{label}{The label for the \code{glassTextInput(...)}}

\item{value}{The current text in the input field in
\code{glassTextInput(...)}}

\item{accept}{The acceptable interval for \code{value} to be in.}

\item{disabled}{A \code{logical} value.}

\item{color}{Color variant: "green" or "purple".}
}
\description{
Update Glass Numeric Input
}
