% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glassRadioButtons.R
\name{updateGlassRadio}
\alias{updateGlassRadio}
\title{Update Glass Radio Group}
\usage{
updateGlassRadio(
  session,
  inputId,
  choices = NULL,
  selected = NULL,
  color = NULL,
  disabled = NULL
)
}
\arguments{
\item{session}{The Shiny session object.}

\item{inputId}{The id of the input object (use simple ID).}

\item{choices}{Optional. New list of choices.}

\item{selected}{Optional. New selection.}

\item{color}{Optional. Color variant: "green" or "purple".}

\item{disabled}{Optional. TRUE to disable.}
}
\description{
Update Glass Radio Group
}
