% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glassSelectizeInput.R
\name{updateGlassSelectizeInput}
\alias{updateGlassSelectizeInput}
\title{Update Glass Selectize Input}
\usage{
updateGlassSelectizeInput(
  session,
  inputId,
  choices = NULL,
  selected = NULL,
  placeholder = NULL
)
}
\arguments{
\item{session}{The Shiny session object.}

\item{inputId}{The id of the selectize input to update.}

\item{choices}{Optional. New list of choices (can be a named vector).}

\item{selected}{Optional. The value(s) to select.}

\item{placeholder}{Optional. New placeholder text.}
}
\description{
Update a glass selectize input from the server.
}
