% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glassLayout.R
\name{updateGlassSidebarHighlight}
\alias{updateGlassSidebarHighlight}
\title{Update Glass Sidebar Highlight}
\usage{
updateGlassSidebarHighlight(session, tabName, enable = TRUE)
}
\arguments{
\item{session}{The Shiny session object.}

\item{tabName}{The `tabName` (data-target) of the item to highlight.}

\item{enable}{Logical. TRUE to start pulsing, FALSE to stop.}
}
\description{
Sends a custom message to the frontend to toggle the notification pulse
on a specific sidebar item. Useful for drawing user attention to a particular
sidebar tab. For example, when the user should move to a specific section.
This is purerly UX/UI related and does not change any app logic.
}
