% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glassTabsetPanel.R
\name{updateGlassTabsetPanel}
\alias{updateGlassTabsetPanel}
\title{Update Glass Tabset Panel (Switch Tab)}
\usage{
updateGlassTabsetPanel(session, inputId, selected)
}
\arguments{
\item{session}{The shiny session object.}

\item{inputId}{The ID of the tabset panel (ns included).}

\item{selected}{The value of the tab to select.}
}
\description{
Sends a custom message to switch the active tab in a glassTabsetPanel.
}
