% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glassInput.R
\name{updateGlassTextInput}
\alias{updateGlassTextInput}
\title{Update Glass Text Input}
\usage{
updateGlassTextInput(
  session,
  inputId,
  label = NULL,
  value = NULL,
  label_icon = NULL,
  disabled = NULL,
  color = NULL
)
}
\arguments{
\item{session}{A valid shiny session object.}

\item{inputId}{The local inputId (not the full ID)}

\item{label}{The label for the \code{glassTextInput(...)}}

\item{value}{The current text in the input field in
\code{glassTextInput(...)}}

\item{label_icon}{The \code{icon(...)} just before the \code{label}}

\item{disabled}{A \code{logical} value.}

\item{color}{Color variant: "green" or "purple".}
}
\description{
Update Glass Text Input
}
