% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glassAnalysisProgress.R
\name{useGlassAnalysisProgress}
\alias{useGlassAnalysisProgress}
\title{HTML dependency for Glass Analysis Progress}
\usage{
useGlassAnalysisProgress()
}
\value{
A tagList with script and stylesheet tags
}
\description{
Include this in your UI to load the JS and CSS assets.
Uses tags$head since this app relies on addResourcePath rather than
being an installed R package.
}
