% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biovar_ported.R
\name{validate_plotting_inputs}
\alias{validate_plotting_inputs}
\title{Validate shared inputs for trace / density plots.}
\usage{
validate_plotting_inputs(plotting_data, parameter_set)
}
\arguments{
\item{plotting_data}{A \code{list} from
\code{calculate_trace_and_posterior_plotting_data()}.}

\item{parameter_set}{A \code{character} string.}
}
\description{
Validate shared inputs for trace / density plots.
}
\keyword{internal}
